Updated bucket size to be 9 digits in a u64.
This allows over half the bits to contain overflow at the expense of taking up more space.

This can certainly be optimized, but I did this for simplicity.
I feel justified in this because my goal in creating this BigUint struct is to understand
the Schonhage-Strassen algorithm, not to make the fastest BigUint.